+++ src/App.jsx
@@ -2,6 +2,7 @@ import { HashRouter, Routes, Route } from 'react-router-dom'
 import Home from './pages/Home'
 import DeckDetail from './pages/DeckDetail'
 import Review from './pages/Review'
+import Browse from './pages/Browse'
 
 export default function App() {
   return (
@@ -10,6 +11,7 @@ export default function App() {
         <Route path="/" element={<Home />} />
         <Route path="/deck/:id" element={<DeckDetail />} />
         <Route path="/review/:id" element={<Review />} />
+        <Route path="/browse/:id" element={<Browse />} />
       </Routes>
     </HashRouter>
   )